requirements.txt
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
py==1.9.0
pygubu==0.10.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
